N3 — move the engine-neutral services to neural/shared/
Roughly 1,900 of pytorch_hooks' lines were protocol services any local
engine needs verbatim (plan §1.2); importing them engine-from-engine would
write a lie into the package structure and make the reference engine
un-removable. So: encoding, metrics, outputs, mechanisms and layout move to
causalab/neural/shared/, along with three helpers the nnsight engine needs
identically (services.py: TensorBundle/BundlePoint + load_tensors,
resolve_roles, site_identity) and the per-layer hybrid stream table
(streams.py) — the one answer that must never diverge between engines.

pytorch_hooks re-exports the moved modules for one deprecation beat.
Remaining internal imports through the pre-rename shim path (test files the
N1 sed did not match because they never said 'backend') now import the real
homes.

Refactor-only. Gates: 1968 passed, corpus digests zero diff, pre-commit
clean.

Plan: cockpit notes/causalab-nnsight-engine-plan.md §2.4 / §6-N3 (D4).
