+++ tests/neural/engines/pytorch_hooks/test_sites_round1_block.py
@@ -16,14 +16,14 @@ from __future__ import annotations
 import pytest
 import torch
 
-from causalab.neural.pytorch_hooks.layout import (
+from causalab.neural.shared.layout import (
     LayoutError,
     from_contract,
     tap_tensor,
     to_contract,
 )
-from causalab.neural.pytorch_hooks.loading import ModelBundle, load_model
-from causalab.neural.pytorch_hooks.sites import resolve_site
+from causalab.neural.engines.pytorch_hooks.loading import ModelBundle, load_model
+from causalab.neural.engines.pytorch_hooks.sites import resolve_site
 from causalab.protocol.errors import ProtocolError, ValidationError
 from causalab.protocol.plan import COMPONENT_RANK
 from causalab.protocol.registry import component_width
